Remove skip validation setting when inlining constraint
when inlining a constraint it get created at the same time as the
table, so there's never any row to be validated. This makes the `NOT
VALID` option a no-op.

Since it's a no-op when inlined, it also leads to a diff when doing a
schema load from the generated structure.sql, since the `NOT VALID`
gets dropped.

So let's be pro-active about it, and just get rid of it as soon as the
constraint gets inlined.